refactor(common): move record reconstruction from SpillableMapUtils to HoodieAvroUtils
The three convertToHoodieRecordPayload overloads (+ private getPreCombineVal)
build a HoodieAvroRecord from an Avro record: extract key/partition from meta
fields, compute the ordering value, optionally strip meta fields, and
instantiate the payload. Nothing about them is specific to spillable maps --
they were only historically located there, and their sole production caller
is HoodieAvroUtils#createHoodieRecordFromAvro (the dispatcher).

Relocate them next to the dispatcher in HoodieAvroUtils; SpillableMapUtils
now contains only spill I/O (readBytesFromDisk/spillToDisk/computePayloadSize).
Behavior-preserving; one test call site updated.
