+++ hudi-common/src/main/java/org/apache/hudi/common/util/SpillableMapUtils.java
@@ -18,28 +18,15 @@
 
 package org.apache.hudi.common.util;
 
-import org.apache.hudi.common.avro.HoodieAvroUtils;
 import org.apache.hudi.common.fs.SizeAwareDataOutputStream;
-import org.apache.hudi.common.model.HoodieAvroRecord;
-import org.apache.hudi.common.model.HoodieKey;
-import org.apache.hudi.common.model.HoodieOperation;
-import org.apache.hudi.common.model.HoodieRecord;
-import org.apache.hudi.common.model.HoodieRecordPayload;
-import org.apache.hudi.common.schema.HoodieSchema;
 import org.apache.hudi.common.util.collection.BitCaskDiskMap.FileEntry;
-import org.apache.hudi.common.util.collection.Pair;
 import org.apache.hudi.exception.HoodieCorruptedDataException;
 
-import org.apache.avro.Schema;
-import org.apache.avro.generic.GenericData;
-import org.apache.avro.generic.GenericRecord;
 
-import javax.annotation.Nullable;
 
 import java.io.IOException;
 import java.io.RandomAccessFile;
 
-import static org.apache.hudi.common.avro.HoodieAvroUtils.getNullableValAsString;
 import static org.apache.hudi.common.util.BinaryUtil.generateChecksum;
 
 /**
@@ -108,73 +95,4 @@ public class SpillableMapUtils {
     return valueSizeEstimator.sizeEstimate(value);
   }
 
-  /**
-   * Utility method to convert bytes to HoodieRecord using schema and payload class.
-   */
-  public static <R> HoodieRecord<R> convertToHoodieRecordPayload(GenericRecord rec, String payloadClazz, String[] preCombineFields, boolean withOperationField) {
-    return convertToHoodieRecordPayload(rec, payloadClazz, preCombineFields,
-        Pair.of(HoodieRecord.RECORD_KEY_METADATA_FIELD, HoodieRecord.PARTITION_PATH_METADATA_FIELD),
-        withOperationField, Option.empty(), Option.empty());
-  }
-
-  public static <R> HoodieRecord<R> convertToHoodieRecordPayload(GenericRecord record, String payloadClazz,
-                                                                 String[] preCombineFields,
-                                                                 boolean withOperationField,
-                                                                 Option<String> partitionName,
-                                                                 Option<HoodieSchema> schemaWithoutMetaFields) {
-    return convertToHoodieRecordPayload(record, payloadClazz, preCombineFields,
-        Pair.of(HoodieRecord.RECORD_KEY_METADATA_FIELD, HoodieRecord.PARTITION_PATH_METADATA_FIELD),
-        withOperationField, partitionName, schemaWithoutMetaFields);
-  }
-
-  /**
-   * Utility method to convert bytes to HoodieRecord using schema and payload class.
-   */
-  public static <R> HoodieRecord<R> convertToHoodieRecordPayload(GenericRecord record, String payloadClazz,
-                                                                 String[] preCombineFields,
-                                                                 Pair<String, String> recordKeyPartitionPathFieldPair,
-                                                                 boolean withOperationField,
-                                                                 Option<String> partitionName,
-                                                                 Option<HoodieSchema> schemaWithoutMetaFields) {
-    final String recKey = record.get(recordKeyPartitionPathFieldPair.getKey()).toString();
-    final String partitionPath = (partitionName.isPresent() ? partitionName.get() :
-        record.get(recordKeyPartitionPathFieldPair.getRight()).toString());
-
-    Comparable preCombineVal = getPreCombineVal(record, preCombineFields);
-    HoodieOperation operation = withOperationField
-        ? HoodieOperation.fromName(getNullableValAsString(record, HoodieRecord.OPERATION_METADATA_FIELD)) : null;
-
-    if (schemaWithoutMetaFields.isPresent()) {
-      Schema schema = schemaWithoutMetaFields.get().toAvroSchema();
-      GenericRecord recordWithoutMetaFields = new GenericData.Record(schema);
-      for (Schema.Field f : schema.getFields()) {
-        recordWithoutMetaFields.put(f.pos(), record.get(f.name()));
-      }
-      record = recordWithoutMetaFields;
-    }
-
-    HoodieRecord<? extends HoodieRecordPayload> hoodieRecord = new HoodieAvroRecord<>(new HoodieKey(recKey, partitionPath),
-        HoodieRecordUtils.loadPayload(payloadClazz, record, preCombineVal), operation);
-    return (HoodieRecord<R>) hoodieRecord;
-  }
-
-  /**
-   * Returns the preCombine value with given field name.
-   *
-   * @param rec The avro record
-   * @param preCombineFields The preCombine field names
-   * @return the preCombine field value or 0 if the field does not exist in the avro schema
-   */
-  private static Comparable getPreCombineVal(GenericRecord rec, @Nullable String[] preCombineFields) {
-    if (preCombineFields == null) {
-      return OrderingValues.getDefault();
-    }
-    // keep consistent with writer side, using Java type for ordering value, see `DefaultHoodieRecordPayload`.
-    return OrderingValues.create(
-        preCombineFields,
-        field -> {
-          Object orderingValue = HoodieAvroUtils.getNestedFieldVal(rec, field, true, false);
-          return orderingValue == null ? OrderingValues.getDefault() : (Comparable) orderingValue;
-        });
-  }
 }
